Move public classes out of internal package
