+++ libs/cardreader/src/main/java/com/woocommerce/android/cardreader/CardReaderDiscoveryEvents.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.cardreader.internal
+package com.woocommerce.android.cardreader
 
 sealed class CardReaderDiscoveryEvents {
     object NotStarted : CardReaderDiscoveryEvents()
