+++ libs/cardreader/src/main/java/com/woocommerce/android/cardreader/CardReaderStatus.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.cardreader.internal
+package com.woocommerce.android.cardreader
 
 enum class CardReaderStatus {
     NOT_CONNECTED,
